// src/redux/types/types.d.ts
export interface Movie {
  id: string;
  name: string;
  banner: string;
  img_url: string;
  rating: string;
  duration: string;
  year: string;
  storyline: string;
  link: string;
  filter: string;
  categories: string[];
}

export interface Series {
  id: string;
  banner: string;
  image: string;
  name: string;
  storyline: string;
  count: string;
  link1: string;
  link2: string;
  link3: string;
  link4: string;
  link5: string;
  link6?: string;
  link7?: string;
  link8?: string;
  link9?: string;
  link10?: string;
  link11?: string;
  link12?: string;
  link13?: string;
  link14?: string;
  link15?: string;
  link16?: string;
  link17?: string;
  link18?: string;
  link19?: string;
  link20?: string;
}